+++ pinot-broker/src/main/java/org/apache/pinot/broker/broker/BrokerAdminApiApplication.java
@@ -34,7 +34,6 @@ import org.apache.hc.core5.util.Timeout;
 import org.apache.helix.HelixManager;
 import org.apache.pinot.broker.queryquota.QueryQuotaManager;
 import org.apache.pinot.broker.requesthandler.BrokerRequestHandler;
-import org.apache.pinot.broker.requesthandler.CursorRequestHandlerDelegate;
 import org.apache.pinot.broker.routing.BrokerRoutingManager;
 import org.apache.pinot.common.cursors.AbstractResultStore;
 import org.apache.pinot.common.http.PoolingHttpClientConnectionManagerHelper;
@@ -77,8 +76,7 @@ public class BrokerAdminApiApplication extends ResourceConfig {
   public BrokerAdminApiApplication(BrokerRoutingManager routingManager, BrokerRequestHandler brokerRequestHandler,
       BrokerMetrics brokerMetrics, PinotConfiguration brokerConf, SqlQueryExecutor sqlQueryExecutor,
       ServerRoutingStatsManager serverRoutingStatsManager, AccessControlFactory accessFactory,
-      HelixManager helixManager, QueryQuotaManager queryQuotaManager,
-      CursorRequestHandlerDelegate cursorRequestHandlerDelegate, AbstractResultStore resultStore) {
+      HelixManager helixManager, QueryQuotaManager queryQuotaManager, AbstractResultStore resultStore) {
     _brokerResourcePackages = brokerConf.getProperty(CommonConstants.Broker.BROKER_RESOURCE_PACKAGES,
         CommonConstants.Broker.DEFAULT_BROKER_RESOURCE_PACKAGES);
     String[] pkgs = _brokerResourcePackages.split(",");
@@ -119,7 +117,6 @@ public class BrokerAdminApiApplication extends ResourceConfig {
         bind(queryQuotaManager).to(QueryQuotaManager.class);
         bind(accessFactory).to(AccessControlFactory.class);
         bind(startTime).named(BrokerAdminApiApplication.START_TIME);
-        bind(cursorRequestHandlerDelegate).to(CursorRequestHandlerDelegate.class);
         bind(resultStore).to(AbstractResultStore.class);
         bind(brokerConf).to(PinotConfiguration.class);
       }
